JAT-0048 Move all packages to modules
