+++ liferay/teams/frontend_infra/frontend_infra_test_map.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python
 
-from liferay.helpers import create_output_files
-from liferay.jira_constants import Filter
-from liferay.jira_liferay import get_jira_connection
-from liferay.helpers_testmap import *
-from liferay.testmap_jira import get_testmap_connection
+from liferay.utils.file_helpers import create_output_files
+from liferay.utils.jira.jira_constants import Filter
+from liferay.utils.jira.jira_liferay import get_jira_connection
+from liferay.utils.testmap_helpers import *
+from liferay.utils.testmap_jira import get_testmap_connection
 
 BUG_THRESHOLD_TAB = 'JIRA-BUGS'
 BUG_THRESHOLD_TAB_RANGE = BUG_THRESHOLD_TAB + '!B3:L'
