+++ liferay/utils/file_helpers.py
