+++ liferay/utils/jira/jira_liferay.py
@@ -1,7 +1,7 @@
 from jira import JIRA
 
-from liferay.jira_constants import Instance
-from liferay.manageCredentialsCrypto import get_credentials
+from liferay.utils.jira.jira_constants import Instance
+from liferay.utils.manageCredentialsCrypto import get_credentials
 
 
 def get_jira_connection(instance_url=Instance.Jira_URL, instance_type=Instance.Type):
