+++ test/test_frontend_infra_test_map.py
@@ -1,8 +1,8 @@
 import unittest
 
-from liferay.frontend_infra_test_map import *
-from liferay.jira_liferay import get_jira_connection
-from liferay.testmap_jira import get_testmap_connection
+from liferay.teams.frontend_infra.frontend_infra_test_map import *
+from liferay.utils.jira.jira_liferay import get_jira_connection
+from liferay.utils.testmap_jira import get_testmap_connection
 
 
 class EchoTestMapTests(unittest.TestCase):
