+++ test/test_frontend_infranstructure.py
@@ -1,7 +1,7 @@
 import unittest
 
-from liferay.frontend_infrastructure import *
-from liferay.jira_liferay import get_jira_connection
+from liferay.teams.frontend_infra.frontend_infrastructure import *
+from liferay.utils.jira.jira_liferay import get_jira_connection
 
 
 class EchoTestMapTests(unittest.TestCase):
